+++ lib/objD/widgets.dart
@@ -1,3 +1,6 @@
 export './execute.dart';
+export './summon.dart';
 export './setblock.dart';
 export './say.dart';
+export './title.dart';
+export './tellraw.dart';